Simplified cloud provider
