+++ pkg/cloudprovider/aws/packing/packing.go
